+++ Panther/exec/main.cpp
@@ -230,6 +230,7 @@ static auto print_num_context_errors(const panther::Context& context, core::Prin
 		.windowsSubsystem       = static_cast<std::optional<ContextConfig::WindowsSubsystem>>(config.windowsSubsystem),
 		.optMode                = config.optMode,
 		.unreachableMode        = config.unreachableMode,
+		.checkedOptionals       = config.checkedOptionals,
 		.compilerExecutablePath = cmd_args_config.executablePath,
 		.workingDirectory       = cmd_args_config.workingDirectory,
 
@@ -1146,6 +1147,7 @@ static auto run_build_system(const pthr::CmdArgsConfig& cmd_args_config, core::P
 		.windowsSubsystem       = std::nullopt,
 		.optMode                = pir::OptMode::SPEED,
 		.unreachableMode        = ContextConfig::UnreachableMode::PANIC,
+		.checkedOptionals       = true,
 		.compilerExecutablePath = cmd_args_config.executablePath,
 		.workingDirectory       = cmd_args_config.workingDirectory,
 
@@ -1320,6 +1322,7 @@ static auto run_scripting(const pthr::CmdArgsConfig& cmd_args_config, core::Prin
 		.windowsSubsystem       = std::nullopt,
 		.optMode                = pir::OptMode::SPEED,
 		.unreachableMode        = ContextConfig::UnreachableMode::PANIC,
+		.checkedOptionals       = true,
 		.compilerExecutablePath = cmd_args_config.executablePath,
 		.workingDirectory       = cmd_args_config.workingDirectory,